% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{sparkxml}
\alias{sparkxml}
\alias{sparkxml-package}
\title{Read xml data files into Spark DataFrames}
\description{
'sparkhaven' is an extension for sparklyr to read xml data files into Spark DataFrames. It uses Spark package spark-xml to load xml data into Spark DataFrames.
}

